requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
langchain==0.2.7
langchain-community==0.2.7
langchain-experimental==0.0.63
weaviate-client==3.26.2
sentence-transformers==2.7.0
gpt4all==2.8.2
ragas==0.1.10
datasets==2.20.0
pydantic==2.7.4
python-dotenv==1.0.1
pydantic-settings==2.3.2
PyPDF2==3.0.1

